+++ tests/dsp/firmware_update.cpp
@@ -4,7 +4,6 @@
 #include <libpldm/base.h>
 #include <libpldm/firmware_update.h>
 #include <libpldm/pldm_types.h>
-#include <libpldm/utils.h>
 
 #include <algorithm>
 #include <array>